requirements.txt
Flask>=3.0.0
flask_sqlalchemy
flask_migrate
pymysql
python-dotenv
jinja2
cryptography
flask-login
debugpy
flask-wtf
flask-security
flask-principal